+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @protohiro/effects-demo
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @protohiro/effects@0.3.0
+
 ## 0.1.2
 
 ### Patch Changes